+++ pkg/proxy/types.go
@@ -18,11 +18,9 @@ package proxy
 
 import (
 	"fmt"
-	"net"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/kubernetes/pkg/proxy/config"
 )
 
@@ -57,75 +55,6 @@ func fmtPortName(in string) string {
 		return ""
 	}
 	return fmt.Sprintf(":%s", in)
-}
-
-// ServicePort is an interface which abstracts information about a service.
-type ServicePort interface {
-	// String returns service string.  An example format can be: `IP:Port/Protocol`.
-	String() string
-	// ClusterIP returns service cluster IP in net.IP format.
-	ClusterIP() net.IP
-	// Port returns service port if present. If return 0 means not present.
-	Port() int
-	// SessionAffinityType returns service session affinity type
-	SessionAffinityType() v1.ServiceAffinity
-	// StickyMaxAgeSeconds returns service max connection age
-	StickyMaxAgeSeconds() int
-	// ExternalIPStrings returns service ExternalIPs as a string array.
-	ExternalIPStrings() []string
-	// LoadBalancerVIPStrings returns service LoadBalancerIPs which are VIP mode as a string array.
-	LoadBalancerVIPStrings() []string
-	// Protocol returns service protocol.
-	Protocol() v1.Protocol
-	// LoadBalancerSourceRanges returns service LoadBalancerSourceRanges if present empty array if not
-	LoadBalancerSourceRanges() []string
-	// HealthCheckNodePort returns service health check node port if present.  If return 0, it means not present.
-	HealthCheckNodePort() int
-	// NodePort returns a service Node port if present. If return 0, it means not present.
-	NodePort() int
-	// ExternalPolicyLocal returns if a service has only node local endpoints for external traffic.
-	ExternalPolicyLocal() bool
-	// InternalPolicyLocal returns if a service has only node local endpoints for internal traffic.
-	InternalPolicyLocal() bool
-	// HintsAnnotation returns the value of the v1.DeprecatedAnnotationTopologyAwareHints annotation.
-	HintsAnnotation() string
-	// ExternallyAccessible returns true if the service port is reachable via something
-	// other than ClusterIP (NodePort/ExternalIP/LoadBalancer)
-	ExternallyAccessible() bool
-	// UsesClusterEndpoints returns true if the service port ever sends traffic to
-	// endpoints based on "Cluster" traffic policy
-	UsesClusterEndpoints() bool
-	// UsesLocalEndpoints returns true if the service port ever sends traffic to
-	// endpoints based on "Local" traffic policy
-	UsesLocalEndpoints() bool
-}
-
-// Endpoint in an interface which abstracts information about an endpoint.
-// TODO: Rename functions to be consistent with ServicePort.
-type Endpoint interface {
-	// String returns endpoint string.  An example format can be: `IP:Port`.
-	// We take the returned value as ServiceEndpoint.Endpoint.
-	String() string
-	// IP returns IP part of the endpoint.
-	IP() string
-	// Port returns the Port part of the endpoint.
-	Port() int
-
-	// IsLocal returns true if the endpoint is running on the same host as kube-proxy.
-	IsLocal() bool
-	// IsReady returns true if an endpoint is ready and not terminating, or
-	// if PublishNotReadyAddresses is set on the service.
-	IsReady() bool
-	// IsServing returns true if an endpoint is ready. It does not account
-	// for terminating state.
-	IsServing() bool
-	// IsTerminating returns true if an endpoint is terminating. For pods,
-	// that is any pod with a deletion timestamp.
-	IsTerminating() bool
-
-	// ZoneHints returns the zone hint for the endpoint. This is based on
-	// endpoint.hints.forZones[0].name in the EndpointSlice API.
-	ZoneHints() sets.Set[string]
 }
 
 // ServiceEndpoint is used to identify a service and one of its endpoint pair.
